refactor(profile): extract shared helpers + ActorProfile, thin resolver (#689)
WO1: Split monolithic [handle]/page.tsx (481 lines) into:
- lib/types.ts — shared ProfileData, IdentityInfo, ViewerContext types
- lib/profile-data.ts — all data-fetching helpers (getProfile, getViewerDid, etc.)
- lib/profile-utils.ts — display helpers (getScopeEmoji, buildServiceUrl, etc.)
- components/ScopeHeader — avatar, name, handle, scope/tier badges
- components/ProfileStats — followers/following/connections + FollowButton
- components/ContactCard — email + phone card
- components/ServiceLinks — Ask/Links/Coffee action buttons (uses buildPublicUrl)
- components/GatedProfile — trust-gated locked view
- components/profiles/ActorProfile — full actor view composing shared components
- [handle]/page.tsx — thin resolver (39 lines, scope switch)

No behavior changes. All profiles render identically.

Refs #689
